build linux images on darwin
The assert in pkgs.nix currently just throws if you try to build on
darwin. As there's no such thing as darwin in a container, I believe
it should be safe to assume that people want to build linux containers.

This defaults to the current platform, but still lets users specify
system to i.e. build x86_64-linux containers from an aarch64-darwin host.